refactor code: refactor instructions folder
