--- programs/maius-payment/src/instructions/merchant.rs
@@ -1,1 +1,0 @@
-use anchor_lang::prelude::*;
