+++ programs/maius-payment/src/instructions/mod.rs
@@ -1,2 +1,14 @@
-pub mod merchant;
+pub mod initialize_merchant;
+pub mod update_merchant;
+pub mod transfer_a_to_b;
+pub mod create_service;
+pub mod initialize_invoice;
+pub mod update_invoice;
+
+pub use initialize_merchant::*;
+pub use update_merchant::*;
+pub use transfer_a_to_b::*;
+pub use create_service::*;
+pub use initialize_invoice::*;
+pub use update_invoice::*;
 // pub mod transaction;
